sacco-backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
